had to move functions and declarations to different files to get the desired effect when expanding or collapsing a field hovering has to be redone because of new structure
